Mistakable German translation (free shipping)
The translation of "help_shipping_item_free_shipping" was unclear; in
fact it suggested that a checked checkbox would switch *on* the shipping
costs for ghe current item - while in fact it makes the shipping free,
i.e. switches coss *off* (right?)

Thus, my suggestion for the German help text makes very clear what to
choose - check the checkbox to make the shipping free of costs.

If I'm wrong, then the English text should be enhanced as well.