context.default_page instead of container.default_page
Drop the aq_parent step when computing the default page.
In Plone 4.3.15 and with LinguaPlone 4.2.1 installed, the former code
tried to get the 'portal_languages' tool in the Zope root (instead of
the Plone site root) when e.g. the @@shop_controlpanel was opened in the
site root.

In any case, in this regard it looks more plausible to me
to inspect the context than the container.